nohabitsfortoday ui

+++ client/src/pages/Tools/HabitTracker/hooks/useTimeFormatState.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 
 export const useTimeFormatState = () => {
-  const [timeFormat, setTimeFormat] = useState<'long' | 'numeric'>('numeric');
+  const [timeFormat, setTimeFormat] = useState<'long' | 'numeric'>('long');
 
   const toggleTimeFormat = () => {
     setTimeFormat((prevTimeFormat) => (prevTimeFormat === 'numeric' ? 'long' : 'numeric'));
